requirements.txt
Flask==2.3.3
googletrans==3.1.0a0
gtts==2.3.2
openai-whisper==20231117
torch